requirements.txt
python-dotenv~=0.21.0
SQLAlchemy~=1.4.45
uvicorn~=0.20.0
fastapi~=0.88.0
pydantic~=1.10.3
PyJWT~=2.6.0
starlette~=0.22.0